backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.0
reportlab==4.0.7
python-dotenv==1.0.0
Pillow==10.1.0